# Improve Crossgen2 compiler perf measurement