Capture the full execution of Crossgen2 with ETW events by measuring command line parsing and moving the point we begin `Loading` measurement to the start of `Program.Run`. This accounts for about 300ms of time we were not measuring previously.

* Add ETW start/stop events around the command line parsing so we can time it and have a number to improve
* Fix up r2rtest `--measure-perf` switch which I broke when I renamed the graph provider and also add the new event.
* Move easily factorable pieces of `Run` into subroutines.

On small compile inputs such as a "Hello World" console app, a surprising amount of time is spent parsing the input arguments in the `System.CommandLine` library. Example perf of helloworld CG2 compilation:
```
Command line processing time: 192.46 ms
Total average compilation time: 594.46 ms
Phase breakdown (average):
    Loading time: 350.26 ms
    Graph processing time: 136.67 ms
        Added 108 nodes to mark stack
        Dependency analysis time: 57.38 ms
        Wall clock JIT time: 71.09 ms
        Total JIT time: 136.93 ms (sum of all threads)
        2 methods JITed
    Emitting time: 103.80 ms
```